SSH2Stream: Allow client to negotiate host key types it will accept
There is no way to adjust what host keys the client will advertise, so
if a server has an ssh-rsa and ssh-dss key but the client only has the
ssh-dss key, verification fails on the client.

Add a hostKeyTypes option to the constructor for clients.  If specified,
this is the list of host key types the client offers that it will accept
in the KEX negotiation.  If not specified, use the old default:
"ssh-rsa,ssh-dss"

Added test-hostkey to exercise the permutations in the negotiation.
This exercises the client and server side of the negotiation.